finish migration to scss modules and make further style updates
