+++ package.json
@@ -20,11 +20,11 @@
     "@types/react-dom": "18.2.7",
     "@types/three": "^0.163.0",
     "assert": "^2.0.0",
-    "eslint": "8.45.0",
-    "eslint-config-next": "13.4.12",
+    "eslint": "9.19.0",
+    "eslint-config-next": "15.1.6",
     "framer-motion": "^10.16.16",
     "fs-extra": "^11.1.1",
-    "next": "14.0.4",
+    "next": "14.2.23",
     "raw-loader": "^4.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
@@ -34,7 +34,5 @@
     "three": "^0.163.0",
     "typescript": "5.3.3"
   },
-  "devDependencies": {
-    "@types/styled-components": "^5.1.26"
-  }
+  "packageManager": "pnpm@10.2.0"
 }
