+++ src/_components/Blocks/ImageBlock/styles.module.scss
@@ -1,6 +1,7 @@
-import styled from 'styled-components'
+@use 'src/_styles/breakpoints' as breakpoints;
+@use 'src/_styles/functions' as *;
 
-export const _ImageBlock = styled.div`
+.ImageBlock {
   position: relative;
   width: 100%;
   height: auto;
@@ -10,13 +11,13 @@ export const _ImageBlock = styled.div`
     width: 100%;
     padding-bottom: 75%;
 
-    ${(props) => props.theme.mediaQueries.tablet} {
+    @include breakpoints.tablet{
       padding-bottom: 56.25%;
     }
   }
-`
+}
 
-export const _Image = styled.img`
+.Image {
   position: absolute;
   top: 0;
   left: 0;
@@ -24,4 +25,4 @@ export const _Image = styled.img`
   height: 100%;
   z-index: 1;
   object-fit: cover;
-`
+}
