--- src/_components/Blocks/PageSections/styles.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const _Sections = styled.div``
-
-export const _Section = styled.section`
-  width: 100%;
-`
