+++ src/_components/Blocks/TextBlock/index.tsx
@@ -1,14 +1,15 @@
-import {
-  _TextBlock,
-    _TextContent,
-} from './styles'
+// Styles
+import styles from "./styles.module.scss";
 
-const TextBlock = ({content}:{content:string}) => {
+const TextBlock = ({ content }: { content: string }) => {
   return (
-    <_TextBlock>
-      <_TextContent dangerouslySetInnerHTML={{__html: content}} />
-    </_TextBlock>
-  )
-}
+    <div className={styles.TextBlock}>
+      <div
+        className={styles.TextContent}
+        dangerouslySetInnerHTML={{ __html: content }}
+      />
+    </div>
+  );
+};
 
-export default TextBlock
+export default TextBlock;
