+++ src/_components/globals/Footer/styles.module.scss
@@ -3,7 +3,7 @@
 
 .FooterWrapper {
   width: 100%;
-  padding: 2rem;
+  padding: 1rem;
   display: flex;
   align-items: center;
   background-color: var(--color-brand_1);
