+++ src/_components/globals/PageHero/styles.module.scss
@@ -1,7 +1,6 @@
 // import { _VideoWrapper } from 'src/_components/Blocks/VideoBlock/styles'
 @use 'src/_styles/breakpoints' as breakpoints;
 @use 'src/_styles/functions' as *;
-@use 'src/_components/Blocks/VideoBlock/styles.module.scss';
 
 .PageHero {
   position: relative;
@@ -27,14 +26,6 @@
   height: 100%;
   z-index: 1;
   background: var(--color-brand_2);
-
-  .videoBlock {
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-  }
 }
 
 .HeroImage {
@@ -47,7 +38,7 @@
 }
 
 .PageTitle {
-  font-size: 6rem;
+  font-size: 5.2rem;
   text-transform: uppercase;
   position: relative;
   z-index: 2;
@@ -55,6 +46,8 @@
   line-height: 1;
   letter-spacing: -0.2rem;
   color: var(--color-white);
+  max-width: 100%;
+  text-shadow: 0.2rem 0.2rem 2rem rgba(var(--rgb-brand_2), 0.4);
 
   @include breakpoints.tablet {
     font-size: 16rem;
