+++ src/_components/globals/PageWrapper/styles.module.scss
@@ -5,5 +5,4 @@
 
 .MainContentWrapper {
   width: 100%;
-  /* overflow: hidden;p */
 }
