+++ src/_components/ui/WorkCard/styles.module.scss
@@ -1,9 +1,8 @@
 @use 'src/_styles/breakpoints' as breakpoints;
-@use 'src/_styles/mixins' as mixins;
 @use 'src/_styles/functions' as *;
 
 .WorkTitle {
-  font-size: 5.5rem;
+  font-size: 4rem;
   letter-spacing: -0.1rem;
   text-transform: uppercase;
   color: var(--color-brand_2);
@@ -42,7 +41,7 @@
   }
 
   h3 {
-    font-size: 2.8rem;
+    font-size: 2rem;
 
     @include breakpoints.desktopSm {
       font-size: RelativeSize(20);
@@ -86,6 +85,7 @@
 .WorkCard {
   display: inline-block;
   width: auto;
+  max-width: 100%;
   position: relative;
   overflow: hidden;
   padding-left: 2.8rem;
