+++ src/_styles/functions.scss
@@ -1,3 +1,17 @@
+@use 'sass:color';
+
+// ///////////////////////////////////////////////////
+// Convert hex colors to rgb
+// ///////////////////////////////////////////////////
+
+@function hexToRGB($hex) {
+  $red: color.channel($hex, "red", $space: rgb);
+  $green: color.channel($hex, "green", $space: rgb);
+  $blue: color.channel($hex, "blue", $space: rgb);
+  
+  @return $red, $green, $blue;
+}
+
 // ///////////////////////////////////////////////////
 // Relative Size
 // ///////////////////////////////////////////////////
