+++ src/_templates/DynamicTemplate/styles.module.scss
@@ -1,0 +1,9 @@
+.DynamicPageContent {
+  width: 100%;
+}
+
+.DynamicPageDetails {
+  width: 100%;
+  height: auto;
+  margin: 0 auto;
+}
