--- src/_templates/DynamicTemplate/styles.ts
@@ -1,11 +1,0 @@
-import { styled } from 'styled-components'
-
-export const _DynamicPageContent = styled.div`
-  width: 100%;
-`
-
-export const _DynamicPageDetails = styled.div`  
-  width: 100%;
-  height: auto;
-  margin: 0 auto;
-`
