+++ src/_templates/WorkTemplate/styles.module.scss
@@ -1,0 +1,19 @@
+@use 'src/_styles/breakpoints' as breakpoints;
+@use 'src/_styles/functions' as *;
+
+.WorkPageContent {
+  width: 100%;
+}
+
+.WorkItems {
+  width: 100%;
+  max-width: 50rem;
+  height: auto;
+  padding: 10rem 1rem 5rem 1rem;
+
+  @include breakpoints.desktopSm {
+    max-width: initial;
+    width: RelativeSize(700);
+    padding: RelativeSize(70) RelativeSize(20);
+  }
+}
